+++ src/components/workspace/connector_type.py
@@ -2,13 +2,14 @@
 
 class ConnectorType:
     # Requires at least typing of Matrix and Scalar
-    Matrix = 0b10000001  # requires matrix type
-    Scalar = 0b10000010  # requires scalar type
-    String = 0b00000100  # requires text/string type
-    Float  = 0b00001000  # requires float type
-    Int    = 0b00010000  # requires integer type
-    Bool   = 0b00100000  # requires boolean type (true/false)
-    Any    = 0b11000000  # any type
+    Matrix   = 0b000000001  # requires matrix type (wrapper type)
+    Scalar   = 0b000000010  # requires scalar type (wrapper type)
+    String   = 0b000000100  # requires text/string type
+    Float    = 0b000001000  # requires float type
+    Int      = 0b000010000  # requires integer type
+    Bool     = 0b000100000  # requires boolean type (true/false)
+    Any      = 0b001000000  # any type
+    Optional = 0b010000000  # nullable (wrapper type)
 
 
 # Matrix: A Square Design
@@ -19,6 +20,7 @@ class ConnectorType:
 # Bool: Red Color
 # Any (Discrete): Black Color
 # Any (Matrix|Scalar): Black Triangle
+# Optional: Black Circle (for now)
 
 # Typing Hierarchical
 # -- every type can only be compatible by itself --
@@ -28,3 +30,4 @@ class ConnectorType:
 # Scalar | Any [= ( Scalar | ( Any | String | Float | Int | Bool ) )
 # Matrix | Int [= Matrix | ( Int | Bool )
 # Scalar | Int [= Matrix | ( Int | Bool )
+# Optional | ( [Any kind of type] ) [= ( Optional | ( [Any kind of type] ) ) | ( [Any kind of type] )