+++ src/constants.py
@@ -5,4 +5,5 @@ TG_INPUT = "inp"  # tag input
 TG_OUTPUT = "out"  # tag output
 FLD_CONST = "const"  # const field
 
+
 DEBUG__ = True